Implement nether claims support for map integrations
