+++ src/main/resources/plugin.yml
@@ -1,5 +1,5 @@
 name: Bonfire
-version: '1.5.3'
+version: '1.6'
 main: com.foenichs.bonfire.Bonfire
 api-version: '26.1.2'
 prefix: Bonfire
